Update README.md
Gochuumon wa usagi desu ka? commonly known as Gochiusa or in its english title "Is the order a rabbit?" is a manga and anime series made by Koi, it is a slice of life where the story revolves around the main characters Cocoa and Chino as they work in Chino's cafe called Rabbit house.